refactor(stats): update naming
